+++ src/rand.h
@@ -1,12 +1,12 @@
 #ifndef SHISHUA_RAND_H
 #define SHISHUA_RAND_H
 
-#include "shishua.h"
+#include <stddef.h>
 #include <stdint.h>
 
-struct rand_chunk{
+struct rand_chunk {
     size_t count;
-    uint8_t buf[128]; //NOTE: must be multiple of 128
+    uint8_t buf[128];  // NOTE: must be multiple of 128
 };
 
 uint8_t randnum_u8(struct rand_chunk *randc, uint8_t min, uint8_t max);
